move methods onto superclass
